Port to ARM/ARM64
