+++ frida-gum/src/instruction_writer/aarch64/register.rs
@@ -1,0 +1,307 @@
+use frida_gum_sys as gum_sys;
+
+#[derive(FromPrimitive, PartialEq, Eq, Clone, Copy, Debug)]
+#[repr(u32)]
+pub enum Aarch64Register {
+    Ffr = gum_sys::arm64_reg_ARM64_REG_FFR as u32,
+    Fp = gum_sys::arm64_reg_ARM64_REG_FP as u32,
+    Lr = gum_sys::arm64_reg_ARM64_REG_LR as u32,
+    Nzcv = gum_sys::arm64_reg_ARM64_REG_NZCV as u32,
+    Sp = gum_sys::arm64_reg_ARM64_REG_SP as u32,
+    Wsp = gum_sys::arm64_reg_ARM64_REG_WSP as u32,
+    Wzr = gum_sys::arm64_reg_ARM64_REG_WZR as u32,
+    Xzr = gum_sys::arm64_reg_ARM64_REG_XZR as u32,
+
+    X0 = gum_sys::arm64_reg_ARM64_REG_X0 as u32,
+    X1 = gum_sys::arm64_reg_ARM64_REG_X1 as u32,
+    X2 = gum_sys::arm64_reg_ARM64_REG_X2 as u32,
+    X3 = gum_sys::arm64_reg_ARM64_REG_X3 as u32,
+    X4 = gum_sys::arm64_reg_ARM64_REG_X4 as u32,
+    X5 = gum_sys::arm64_reg_ARM64_REG_X5 as u32,
+    X6 = gum_sys::arm64_reg_ARM64_REG_X6 as u32,
+    X7 = gum_sys::arm64_reg_ARM64_REG_X7 as u32,
+    X8 = gum_sys::arm64_reg_ARM64_REG_X8 as u32,
+    X9 = gum_sys::arm64_reg_ARM64_REG_X9 as u32,
+    X10 = gum_sys::arm64_reg_ARM64_REG_X10 as u32,
+    X11 = gum_sys::arm64_reg_ARM64_REG_X11 as u32,
+    X12 = gum_sys::arm64_reg_ARM64_REG_X12 as u32,
+    X13 = gum_sys::arm64_reg_ARM64_REG_X13 as u32,
+    X14 = gum_sys::arm64_reg_ARM64_REG_X14 as u32,
+    X15 = gum_sys::arm64_reg_ARM64_REG_X15 as u32,
+    X16 = gum_sys::arm64_reg_ARM64_REG_X16 as u32,
+    X17 = gum_sys::arm64_reg_ARM64_REG_X17 as u32,
+    X18 = gum_sys::arm64_reg_ARM64_REG_X18 as u32,
+    X19 = gum_sys::arm64_reg_ARM64_REG_X19 as u32,
+    X20 = gum_sys::arm64_reg_ARM64_REG_X20 as u32,
+    X21 = gum_sys::arm64_reg_ARM64_REG_X21 as u32,
+    X22 = gum_sys::arm64_reg_ARM64_REG_X22 as u32,
+    X23 = gum_sys::arm64_reg_ARM64_REG_X23 as u32,
+    X24 = gum_sys::arm64_reg_ARM64_REG_X24 as u32,
+    X25 = gum_sys::arm64_reg_ARM64_REG_X25 as u32,
+    X26 = gum_sys::arm64_reg_ARM64_REG_X26 as u32,
+    X27 = gum_sys::arm64_reg_ARM64_REG_X27 as u32,
+    X28 = gum_sys::arm64_reg_ARM64_REG_X28 as u32,
+
+    W0 = gum_sys::arm64_reg_ARM64_REG_W0 as u32,
+    W1 = gum_sys::arm64_reg_ARM64_REG_W1 as u32,
+    W2 = gum_sys::arm64_reg_ARM64_REG_W2 as u32,
+    W3 = gum_sys::arm64_reg_ARM64_REG_W3 as u32,
+    W4 = gum_sys::arm64_reg_ARM64_REG_W4 as u32,
+    W5 = gum_sys::arm64_reg_ARM64_REG_W5 as u32,
+    W6 = gum_sys::arm64_reg_ARM64_REG_W6 as u32,
+    W7 = gum_sys::arm64_reg_ARM64_REG_W7 as u32,
+    W8 = gum_sys::arm64_reg_ARM64_REG_W8 as u32,
+    W9 = gum_sys::arm64_reg_ARM64_REG_W9 as u32,
+    W10 = gum_sys::arm64_reg_ARM64_REG_W10 as u32,
+    W11 = gum_sys::arm64_reg_ARM64_REG_W11 as u32,
+    W12 = gum_sys::arm64_reg_ARM64_REG_W12 as u32,
+    W13 = gum_sys::arm64_reg_ARM64_REG_W13 as u32,
+    W14 = gum_sys::arm64_reg_ARM64_REG_W14 as u32,
+    W15 = gum_sys::arm64_reg_ARM64_REG_W15 as u32,
+    W16 = gum_sys::arm64_reg_ARM64_REG_W16 as u32,
+    W17 = gum_sys::arm64_reg_ARM64_REG_W17 as u32,
+    W18 = gum_sys::arm64_reg_ARM64_REG_W18 as u32,
+    W19 = gum_sys::arm64_reg_ARM64_REG_W19 as u32,
+    W20 = gum_sys::arm64_reg_ARM64_REG_W20 as u32,
+    W21 = gum_sys::arm64_reg_ARM64_REG_W21 as u32,
+    W22 = gum_sys::arm64_reg_ARM64_REG_W22 as u32,
+    W23 = gum_sys::arm64_reg_ARM64_REG_W23 as u32,
+    W24 = gum_sys::arm64_reg_ARM64_REG_W24 as u32,
+    W25 = gum_sys::arm64_reg_ARM64_REG_W25 as u32,
+    W26 = gum_sys::arm64_reg_ARM64_REG_W26 as u32,
+    W27 = gum_sys::arm64_reg_ARM64_REG_W27 as u32,
+    W28 = gum_sys::arm64_reg_ARM64_REG_W28 as u32,
+    W29 = gum_sys::arm64_reg_ARM64_REG_W29 as u32,
+    W30 = gum_sys::arm64_reg_ARM64_REG_W30 as u32,
+
+    S0 = gum_sys::arm64_reg_ARM64_REG_S0 as u32,
+    S1 = gum_sys::arm64_reg_ARM64_REG_S1 as u32,
+    S2 = gum_sys::arm64_reg_ARM64_REG_S2 as u32,
+    S3 = gum_sys::arm64_reg_ARM64_REG_S3 as u32,
+    S4 = gum_sys::arm64_reg_ARM64_REG_S4 as u32,
+    S5 = gum_sys::arm64_reg_ARM64_REG_S5 as u32,
+    S6 = gum_sys::arm64_reg_ARM64_REG_S6 as u32,
+    S7 = gum_sys::arm64_reg_ARM64_REG_S7 as u32,
+    S8 = gum_sys::arm64_reg_ARM64_REG_S8 as u32,
+    S9 = gum_sys::arm64_reg_ARM64_REG_S9 as u32,
+    S10 = gum_sys::arm64_reg_ARM64_REG_S10 as u32,
+    S11 = gum_sys::arm64_reg_ARM64_REG_S11 as u32,
+    S12 = gum_sys::arm64_reg_ARM64_REG_S12 as u32,
+    S13 = gum_sys::arm64_reg_ARM64_REG_S13 as u32,
+    S14 = gum_sys::arm64_reg_ARM64_REG_S14 as u32,
+    S15 = gum_sys::arm64_reg_ARM64_REG_S15 as u32,
+    S16 = gum_sys::arm64_reg_ARM64_REG_S16 as u32,
+    S17 = gum_sys::arm64_reg_ARM64_REG_S17 as u32,
+    S18 = gum_sys::arm64_reg_ARM64_REG_S18 as u32,
+    S19 = gum_sys::arm64_reg_ARM64_REG_S19 as u32,
+    S20 = gum_sys::arm64_reg_ARM64_REG_S20 as u32,
+    S21 = gum_sys::arm64_reg_ARM64_REG_S21 as u32,
+    S22 = gum_sys::arm64_reg_ARM64_REG_S22 as u32,
+    S23 = gum_sys::arm64_reg_ARM64_REG_S23 as u32,
+    S24 = gum_sys::arm64_reg_ARM64_REG_S24 as u32,
+    S25 = gum_sys::arm64_reg_ARM64_REG_S25 as u32,
+    S26 = gum_sys::arm64_reg_ARM64_REG_S26 as u32,
+    S27 = gum_sys::arm64_reg_ARM64_REG_S27 as u32,
+    S28 = gum_sys::arm64_reg_ARM64_REG_S28 as u32,
+    S29 = gum_sys::arm64_reg_ARM64_REG_S29 as u32,
+    S30 = gum_sys::arm64_reg_ARM64_REG_S30 as u32,
+    S31 = gum_sys::arm64_reg_ARM64_REG_S31 as u32,
+
+    H0 = gum_sys::arm64_reg_ARM64_REG_H0 as u32,
+    H1 = gum_sys::arm64_reg_ARM64_REG_H1 as u32,
+    H2 = gum_sys::arm64_reg_ARM64_REG_H2 as u32,
+    H3 = gum_sys::arm64_reg_ARM64_REG_H3 as u32,
+    H4 = gum_sys::arm64_reg_ARM64_REG_H4 as u32,
+    H5 = gum_sys::arm64_reg_ARM64_REG_H5 as u32,
+    H6 = gum_sys::arm64_reg_ARM64_REG_H6 as u32,
+    H7 = gum_sys::arm64_reg_ARM64_REG_H7 as u32,
+    H8 = gum_sys::arm64_reg_ARM64_REG_H8 as u32,
+    H9 = gum_sys::arm64_reg_ARM64_REG_H9 as u32,
+    H10 = gum_sys::arm64_reg_ARM64_REG_H10 as u32,
+    H11 = gum_sys::arm64_reg_ARM64_REG_H11 as u32,
+    H12 = gum_sys::arm64_reg_ARM64_REG_H12 as u32,
+    H13 = gum_sys::arm64_reg_ARM64_REG_H13 as u32,
+    H14 = gum_sys::arm64_reg_ARM64_REG_H14 as u32,
+    H15 = gum_sys::arm64_reg_ARM64_REG_H15 as u32,
+    H16 = gum_sys::arm64_reg_ARM64_REG_H16 as u32,
+    H17 = gum_sys::arm64_reg_ARM64_REG_H17 as u32,
+    H18 = gum_sys::arm64_reg_ARM64_REG_H18 as u32,
+    H19 = gum_sys::arm64_reg_ARM64_REG_H19 as u32,
+    H20 = gum_sys::arm64_reg_ARM64_REG_H20 as u32,
+    H21 = gum_sys::arm64_reg_ARM64_REG_H21 as u32,
+    H22 = gum_sys::arm64_reg_ARM64_REG_H22 as u32,
+    H23 = gum_sys::arm64_reg_ARM64_REG_H23 as u32,
+    H24 = gum_sys::arm64_reg_ARM64_REG_H24 as u32,
+    H25 = gum_sys::arm64_reg_ARM64_REG_H25 as u32,
+    H26 = gum_sys::arm64_reg_ARM64_REG_H26 as u32,
+    H27 = gum_sys::arm64_reg_ARM64_REG_H27 as u32,
+    H28 = gum_sys::arm64_reg_ARM64_REG_H28 as u32,
+    H29 = gum_sys::arm64_reg_ARM64_REG_H29 as u32,
+    H30 = gum_sys::arm64_reg_ARM64_REG_H30 as u32,
+    H31 = gum_sys::arm64_reg_ARM64_REG_H31 as u32,
+
+    B0 = gum_sys::arm64_reg_ARM64_REG_B0 as u32,
+    B1 = gum_sys::arm64_reg_ARM64_REG_B1 as u32,
+    B2 = gum_sys::arm64_reg_ARM64_REG_B2 as u32,
+    B3 = gum_sys::arm64_reg_ARM64_REG_B3 as u32,
+    B4 = gum_sys::arm64_reg_ARM64_REG_B4 as u32,
+    B5 = gum_sys::arm64_reg_ARM64_REG_B5 as u32,
+    B6 = gum_sys::arm64_reg_ARM64_REG_B6 as u32,
+    B7 = gum_sys::arm64_reg_ARM64_REG_B7 as u32,
+    B8 = gum_sys::arm64_reg_ARM64_REG_B8 as u32,
+    B9 = gum_sys::arm64_reg_ARM64_REG_B9 as u32,
+    B10 = gum_sys::arm64_reg_ARM64_REG_B10 as u32,
+    B11 = gum_sys::arm64_reg_ARM64_REG_B11 as u32,
+    B12 = gum_sys::arm64_reg_ARM64_REG_B12 as u32,
+    B13 = gum_sys::arm64_reg_ARM64_REG_B13 as u32,
+    B14 = gum_sys::arm64_reg_ARM64_REG_B14 as u32,
+    B15 = gum_sys::arm64_reg_ARM64_REG_B15 as u32,
+    B16 = gum_sys::arm64_reg_ARM64_REG_B16 as u32,
+    B17 = gum_sys::arm64_reg_ARM64_REG_B17 as u32,
+    B18 = gum_sys::arm64_reg_ARM64_REG_B18 as u32,
+    B19 = gum_sys::arm64_reg_ARM64_REG_B19 as u32,
+    B20 = gum_sys::arm64_reg_ARM64_REG_B20 as u32,
+    B21 = gum_sys::arm64_reg_ARM64_REG_B21 as u32,
+    B22 = gum_sys::arm64_reg_ARM64_REG_B22 as u32,
+    B23 = gum_sys::arm64_reg_ARM64_REG_B23 as u32,
+    B24 = gum_sys::arm64_reg_ARM64_REG_B24 as u32,
+    B25 = gum_sys::arm64_reg_ARM64_REG_B25 as u32,
+    B26 = gum_sys::arm64_reg_ARM64_REG_B26 as u32,
+    B27 = gum_sys::arm64_reg_ARM64_REG_B27 as u32,
+    B28 = gum_sys::arm64_reg_ARM64_REG_B28 as u32,
+    B29 = gum_sys::arm64_reg_ARM64_REG_B29 as u32,
+    B30 = gum_sys::arm64_reg_ARM64_REG_B30 as u32,
+    B31 = gum_sys::arm64_reg_ARM64_REG_B31 as u32,
+
+    D0 = gum_sys::arm64_reg_ARM64_REG_D0 as u32,
+    D1 = gum_sys::arm64_reg_ARM64_REG_D1 as u32,
+    D2 = gum_sys::arm64_reg_ARM64_REG_D2 as u32,
+    D3 = gum_sys::arm64_reg_ARM64_REG_D3 as u32,
+    D4 = gum_sys::arm64_reg_ARM64_REG_D4 as u32,
+    D5 = gum_sys::arm64_reg_ARM64_REG_D5 as u32,
+    D6 = gum_sys::arm64_reg_ARM64_REG_D6 as u32,
+    D7 = gum_sys::arm64_reg_ARM64_REG_D7 as u32,
+    D8 = gum_sys::arm64_reg_ARM64_REG_D8 as u32,
+    D9 = gum_sys::arm64_reg_ARM64_REG_D9 as u32,
+    D10 = gum_sys::arm64_reg_ARM64_REG_D10 as u32,
+    D11 = gum_sys::arm64_reg_ARM64_REG_D11 as u32,
+    D12 = gum_sys::arm64_reg_ARM64_REG_D12 as u32,
+    D13 = gum_sys::arm64_reg_ARM64_REG_D13 as u32,
+    D14 = gum_sys::arm64_reg_ARM64_REG_D14 as u32,
+    D15 = gum_sys::arm64_reg_ARM64_REG_D15 as u32,
+    D16 = gum_sys::arm64_reg_ARM64_REG_D16 as u32,
+    D17 = gum_sys::arm64_reg_ARM64_REG_D17 as u32,
+    D18 = gum_sys::arm64_reg_ARM64_REG_D18 as u32,
+    D19 = gum_sys::arm64_reg_ARM64_REG_D19 as u32,
+    D20 = gum_sys::arm64_reg_ARM64_REG_D20 as u32,
+    D21 = gum_sys::arm64_reg_ARM64_REG_D21 as u32,
+    D22 = gum_sys::arm64_reg_ARM64_REG_D22 as u32,
+    D23 = gum_sys::arm64_reg_ARM64_REG_D23 as u32,
+    D24 = gum_sys::arm64_reg_ARM64_REG_D24 as u32,
+    D25 = gum_sys::arm64_reg_ARM64_REG_D25 as u32,
+    D26 = gum_sys::arm64_reg_ARM64_REG_D26 as u32,
+    D27 = gum_sys::arm64_reg_ARM64_REG_D27 as u32,
+    D28 = gum_sys::arm64_reg_ARM64_REG_D28 as u32,
+    D29 = gum_sys::arm64_reg_ARM64_REG_D29 as u32,
+    D30 = gum_sys::arm64_reg_ARM64_REG_D30 as u32,
+    D31 = gum_sys::arm64_reg_ARM64_REG_D31 as u32,
+
+    Q0 = gum_sys::arm64_reg_ARM64_REG_Q0 as u32,
+    Q1 = gum_sys::arm64_reg_ARM64_REG_Q1 as u32,
+    Q2 = gum_sys::arm64_reg_ARM64_REG_Q2 as u32,
+    Q3 = gum_sys::arm64_reg_ARM64_REG_Q3 as u32,
+    Q4 = gum_sys::arm64_reg_ARM64_REG_Q4 as u32,
+    Q5 = gum_sys::arm64_reg_ARM64_REG_Q5 as u32,
+    Q6 = gum_sys::arm64_reg_ARM64_REG_Q6 as u32,
+    Q7 = gum_sys::arm64_reg_ARM64_REG_Q7 as u32,
+    Q8 = gum_sys::arm64_reg_ARM64_REG_Q8 as u32,
+    Q9 = gum_sys::arm64_reg_ARM64_REG_Q9 as u32,
+    Q10 = gum_sys::arm64_reg_ARM64_REG_Q10 as u32,
+    Q11 = gum_sys::arm64_reg_ARM64_REG_Q11 as u32,
+    Q12 = gum_sys::arm64_reg_ARM64_REG_Q12 as u32,
+    Q13 = gum_sys::arm64_reg_ARM64_REG_Q13 as u32,
+    Q14 = gum_sys::arm64_reg_ARM64_REG_Q14 as u32,
+    Q15 = gum_sys::arm64_reg_ARM64_REG_Q15 as u32,
+    Q16 = gum_sys::arm64_reg_ARM64_REG_Q16 as u32,
+    Q17 = gum_sys::arm64_reg_ARM64_REG_Q17 as u32,
+    Q18 = gum_sys::arm64_reg_ARM64_REG_Q18 as u32,
+    Q19 = gum_sys::arm64_reg_ARM64_REG_Q19 as u32,
+    Q20 = gum_sys::arm64_reg_ARM64_REG_Q20 as u32,
+    Q21 = gum_sys::arm64_reg_ARM64_REG_Q21 as u32,
+    Q22 = gum_sys::arm64_reg_ARM64_REG_Q22 as u32,
+    Q23 = gum_sys::arm64_reg_ARM64_REG_Q23 as u32,
+    Q24 = gum_sys::arm64_reg_ARM64_REG_Q24 as u32,
+    Q25 = gum_sys::arm64_reg_ARM64_REG_Q25 as u32,
+    Q26 = gum_sys::arm64_reg_ARM64_REG_Q26 as u32,
+    Q27 = gum_sys::arm64_reg_ARM64_REG_Q27 as u32,
+    Q28 = gum_sys::arm64_reg_ARM64_REG_Q28 as u32,
+    Q29 = gum_sys::arm64_reg_ARM64_REG_Q29 as u32,
+    Q30 = gum_sys::arm64_reg_ARM64_REG_Q30 as u32,
+    Q31 = gum_sys::arm64_reg_ARM64_REG_Q31 as u32,
+
+    Z0 = gum_sys::arm64_reg_ARM64_REG_Z0 as u32,
+    Z1 = gum_sys::arm64_reg_ARM64_REG_Z1 as u32,
+    Z2 = gum_sys::arm64_reg_ARM64_REG_Z2 as u32,
+    Z3 = gum_sys::arm64_reg_ARM64_REG_Z3 as u32,
+    Z4 = gum_sys::arm64_reg_ARM64_REG_Z4 as u32,
+    Z5 = gum_sys::arm64_reg_ARM64_REG_Z5 as u32,
+    Z6 = gum_sys::arm64_reg_ARM64_REG_Z6 as u32,
+    Z7 = gum_sys::arm64_reg_ARM64_REG_Z7 as u32,
+    Z8 = gum_sys::arm64_reg_ARM64_REG_Z8 as u32,
+    Z9 = gum_sys::arm64_reg_ARM64_REG_Z9 as u32,
+    Z10 = gum_sys::arm64_reg_ARM64_REG_Z10 as u32,
+    Z11 = gum_sys::arm64_reg_ARM64_REG_Z11 as u32,
+    Z12 = gum_sys::arm64_reg_ARM64_REG_Z12 as u32,
+    Z13 = gum_sys::arm64_reg_ARM64_REG_Z13 as u32,
+    Z14 = gum_sys::arm64_reg_ARM64_REG_Z14 as u32,
+    Z15 = gum_sys::arm64_reg_ARM64_REG_Z15 as u32,
+    Z16 = gum_sys::arm64_reg_ARM64_REG_Z16 as u32,
+    Z17 = gum_sys::arm64_reg_ARM64_REG_Z17 as u32,
+    Z18 = gum_sys::arm64_reg_ARM64_REG_Z18 as u32,
+    Z19 = gum_sys::arm64_reg_ARM64_REG_Z19 as u32,
+    Z20 = gum_sys::arm64_reg_ARM64_REG_Z20 as u32,
+    Z21 = gum_sys::arm64_reg_ARM64_REG_Z21 as u32,
+    Z22 = gum_sys::arm64_reg_ARM64_REG_Z22 as u32,
+    Z23 = gum_sys::arm64_reg_ARM64_REG_Z23 as u32,
+    Z24 = gum_sys::arm64_reg_ARM64_REG_Z24 as u32,
+    Z25 = gum_sys::arm64_reg_ARM64_REG_Z25 as u32,
+    Z26 = gum_sys::arm64_reg_ARM64_REG_Z26 as u32,
+    Z27 = gum_sys::arm64_reg_ARM64_REG_Z27 as u32,
+    Z28 = gum_sys::arm64_reg_ARM64_REG_Z28 as u32,
+    Z29 = gum_sys::arm64_reg_ARM64_REG_Z29 as u32,
+    Z30 = gum_sys::arm64_reg_ARM64_REG_Z30 as u32,
+    Z31 = gum_sys::arm64_reg_ARM64_REG_Z31 as u32,
+
+    V0 = gum_sys::arm64_reg_ARM64_REG_V0 as u32,
+    V1 = gum_sys::arm64_reg_ARM64_REG_V1 as u32,
+    V2 = gum_sys::arm64_reg_ARM64_REG_V2 as u32,
+    V3 = gum_sys::arm64_reg_ARM64_REG_V3 as u32,
+    V4 = gum_sys::arm64_reg_ARM64_REG_V4 as u32,
+    V5 = gum_sys::arm64_reg_ARM64_REG_V5 as u32,
+    V6 = gum_sys::arm64_reg_ARM64_REG_V6 as u32,
+    V7 = gum_sys::arm64_reg_ARM64_REG_V7 as u32,
+    V8 = gum_sys::arm64_reg_ARM64_REG_V8 as u32,
+    V9 = gum_sys::arm64_reg_ARM64_REG_V9 as u32,
+    V10 = gum_sys::arm64_reg_ARM64_REG_V10 as u32,
+    V11 = gum_sys::arm64_reg_ARM64_REG_V11 as u32,
+    V12 = gum_sys::arm64_reg_ARM64_REG_V12 as u32,
+    V13 = gum_sys::arm64_reg_ARM64_REG_V13 as u32,
+    V14 = gum_sys::arm64_reg_ARM64_REG_V14 as u32,
+    V15 = gum_sys::arm64_reg_ARM64_REG_V15 as u32,
+    V16 = gum_sys::arm64_reg_ARM64_REG_V16 as u32,
+    V17 = gum_sys::arm64_reg_ARM64_REG_V17 as u32,
+    V18 = gum_sys::arm64_reg_ARM64_REG_V18 as u32,
+    V19 = gum_sys::arm64_reg_ARM64_REG_V19 as u32,
+    V20 = gum_sys::arm64_reg_ARM64_REG_V20 as u32,
+    V21 = gum_sys::arm64_reg_ARM64_REG_V21 as u32,
+    V22 = gum_sys::arm64_reg_ARM64_REG_V22 as u32,
+    V23 = gum_sys::arm64_reg_ARM64_REG_V23 as u32,
+    V24 = gum_sys::arm64_reg_ARM64_REG_V24 as u32,
+    V25 = gum_sys::arm64_reg_ARM64_REG_V25 as u32,
+    V26 = gum_sys::arm64_reg_ARM64_REG_V26 as u32,
+    V27 = gum_sys::arm64_reg_ARM64_REG_V27 as u32,
+    V28 = gum_sys::arm64_reg_ARM64_REG_V28 as u32,
+    V29 = gum_sys::arm64_reg_ARM64_REG_V29 as u32,
+    V30 = gum_sys::arm64_reg_ARM64_REG_V30 as u32,
+    V31 = gum_sys::arm64_reg_ARM64_REG_V31 as u32,
+}
